reorganize Layout components
