+++ frontend/src/components/Layout/Modal.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import { resetModal } from '../app/features/modalSlice';
-import { User } from '../app/features/types'
-import { useAppSelector, useAppDispatch } from '../app/hooks';
-import { handleSendInvite, handleInviteAccepted, handleDeclineInvite, handleSendVideoInvite } from '../services/publishers';
+import { resetModal } from '../../app/features/modalSlice';
+import { User } from '../../app/features/types'
+import { useAppSelector, useAppDispatch } from '../../app/hooks';
+import { handleSendInvite, handleInviteAccepted, handleDeclineInvite, handleSendVideoInvite } from '../../services/publishers';
 
 const ActionModal: React.FC = () => {
   const dispatch = useAppDispatch();
