+++ frontend/src/components/Layout/Notification.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useAppSelector } from '../app/hooks';
+import { useAppSelector } from '../../app/hooks';
 
 
 const Notification = () => {
